Test.php
file linked with the pages of Pre-register (see front-end). Also linked with db_config, where are stored database logs and informations. This file connects to database, see if connection is possible,  take the info of the inputbox, protects against SQL injections, and see if email is already stored in the database, (if not, it write the email in the database). Depending of the case, the php sends back the user to the appropriate html page.

TODO: -Send an automatic email of confirmation
-protect better database_logs
-encrypt data in the database (phyisically if possible)
-make a js script to avois empty fields registering (because for now it is possible)